+++ setup.py
@@ -70,6 +70,8 @@ setup(
         "nbformat",
         "toposort",
         "pulp >=2.0",
+        "smart_open",
+        "filelock",
     ],
     extras_require={
         "reports": ["jinja2", "networkx", "pygments", "pygraphviz"],
