+++ test-environment.yml
@@ -57,3 +57,4 @@ dependencies:
   - peppy
   - eido
   - glpk
+  - smart_open
